// types.ts
export interface Feature {
  icon: React.ReactNode;
  title: string;
  description: string;
}

export interface ShowcaseItem {
  image: string;
  category: string;
  title: string;
  description: string;
}

export interface Testimonial {
  quote: string;
  avatar: string;
  name: string;
  role: string;
  company?: string;
}

export interface AccessOption {
  icon: React.ReactNode;
  title: string;
  description: string;
  ctaText: string;
  ctaLink: string;
}

export interface FooterLink {
    name: string;
    href: string;
}

export interface FooterColumn {
    title: string;
    links: FooterLink[];
}